requirements.txt
django==5.1
djangorestframework==3.15.2
cbor2==5.6.4
pillow==10.4.0
